fix table update when sql fail
The table is updated even if the sql failed so it does not reflect the actual database. I now implemented a conditional to check if the sql statement was successful or not and only update the table if the sql was successful for the change for both the submit/insert and confirm/update operations.